strlen corr

+++ ft_putstr_fd.c
@@ -17,5 +17,5 @@ void	ft_putstr_fd(char *s, int fd)
 {
 	if (!s)
 		return ;
-	write(fd, s, (unsigned int)ft_strlen(s));
+	write(fd, s, ft_strlen(s));
 }
