+++ ft_strlcat.c
@@ -18,8 +18,8 @@ size_t	ft_strlcat(char *dst, char *src, size_t dstsize)
 	size_t	lendst;
 	size_t	lensrc;
 
-	lendst = (size_t)ft_strlen(dst);
-	lensrc = (size_t)ft_strlen(src);
+	lendst = ft_strlen(dst);
+	lensrc = ft_strlen(src);
 	if (dstsize <= lendst)
 		return (dstsize + lensrc);
 	ft_strlcpy(dst + lendst, src, dstsize - lendst);
